lint and test fixes
